chore: add shared type definitions
